new file:   blockchain.cpp 	new file:   blockchain.h 	modified:   main.cpp
	cleaned up the code, added headers for block and blockchain class
