+++ .gitignore
@@ -1,0 +1,3 @@
+.vscode/launch.json
+.vscode/tasks.json
+main
